src/test/java/org/symphonyoss/integration/webhook/jira/parser/V2ParserFactoryTest.java
/**
 * Copyright 2016-2017 Symphony Integrations - Symphony LLC
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package org.symphonyoss.integration.webhook.jira.parser;

import static org.junit.Assert.assertFalse;
import static org.junit.Assert.assertTrue;

import org.apache.commons.lang3.NotImplementedException;
import org.junit.Test;
import org.symphonyoss.integration.model.message.MessageMLVersion;

/**
 * Unit test for {@link V2ParserFactory}
 * Created by rsanchez on 23/03/17.
 */
public class V2ParserFactoryTest {

  private V2ParserFactory factory = new V2ParserFactory();

  @Test
  public void testNotAcceptable() {
    assertFalse(factory.accept(MessageMLVersion.V1));
  }

  @Test
  public void testAcceptable() {
    assertTrue(factory.accept(MessageMLVersion.V2));
  }

  @Test(expected = NotImplementedException.class)
  public void testGetParser() {
    factory.getParser(null);
  }

}
